Removed test case from XMLHandler main method and migrated to tests/test_xml_handler
