+++ python/lsst/iip/tests/.cache/v/cache/lastfailed
@@ -1,0 +1,5 @@
+{
+  "test_ack_sb.py": true, 
+  "test_job_sb.py": true, 
+  "test_xml_handler.py": true
+}
